add safeguards package with injection split
