+++ sdk/src/unplug/safeguards/__init__.py
@@ -1,0 +1,15 @@
+"""Threat-class safeguards (detection layer)."""
+
+from __future__ import annotations
+
+from unplug.safeguards.base import BaseScanner, ModelScanner, RegexScanner, Scanner
+from unplug.safeguards.registry import SafeguardRegistry, ScannerRegistry
+
+__all__ = [
+    "BaseScanner",
+    "ModelScanner",
+    "RegexScanner",
+    "SafeguardRegistry",
+    "Scanner",
+    "ScannerRegistry",
+]
