+++ sdk/src/unplug/safeguards/injection/__init__.py
@@ -1,0 +1,7 @@
+"""Prompt injection safeguard."""
+
+from __future__ import annotations
+
+from unplug.safeguards.injection.scanner import InjectionScanner
+
+__all__ = ["InjectionScanner"]
