+++ sdk/src/unplug/safeguards/injection/patterns.py
@@ -1,0 +1,78 @@
+"""Injection and jailbreak regex patterns."""
+
+from __future__ import annotations
+
+import re
+
+INJECTION_PATTERNS: list[tuple[str, re.Pattern[str]]] = [
+    (
+        "ignore_previous",
+        re.compile(
+            r"(?i)(ignore|forget|disregard|override|bypass)\s+(all\s+)?"
+            r"(previous|prior|above|earlier)\s+(instructions?|prompts?|rules?|guidelines?)",
+        ),
+    ),
+    (
+        "persona_replacement",
+        re.compile(
+            r"(?i)(you\s+are\s+now|act\s+as|pretend\s+(to\s+be|you\s+are)|from\s+now\s+on\s+you)",
+        ),
+    ),
+    (
+        "reveal_prompt",
+        re.compile(
+            r"(?i)(reveal|show|display|print|output|repeat|tell\s+me)\s+(your\s+)?"
+            r"(system\s+)?(prompt|instructions?|rules?|guidelines?|configuration)",
+        ),
+    ),
+    (
+        "system_extraction_alt",
+        re.compile(
+            r"(?i)(what\s+(are|is)\s+your\s+(system\s+)?(prompt|instructions?|rules?)|"
+            r"copy\s+your\s+(initial|system)\s+prompt)",
+        ),
+    ),
+    (
+        "developer_mode",
+        re.compile(
+            r"(?i)(developer\s+mode|DAN\s+mode|jailbreak|do\s+anything\s+now|no\s+restrictions?)",
+        ),
+    ),
+    (
+        "closing_delimiter",
+        re.compile(
+            r"(```|<\/?system>|<\/?instruction>|\[\/INST\]|<\|im_end\|>)",
+        ),
+    ),
+    (
+        "base64_payload",
+        re.compile(
+            r"[A-Za-z0-9+/]{20,}={0,2}",
+        ),
+    ),
+    (
+        "hex_encoding",
+        re.compile(
+            r"(?i)(?:\\x[0-9a-f]{2}){4,}",
+        ),
+    ),
+    (
+        "unicode_escape",
+        re.compile(
+            r"(?:\\u[0-9a-fA-F]{4}){3,}",
+        ),
+    ),
+    (
+        "indirect_instruction",
+        re.compile(
+            r"(?i)(when\s+you\s+read\s+this|important\s+instruction|hidden\s+instruction|"
+            r"ignore\s+the\s+document\s+and|do\s+not\s+follow\s+the\s+user)",
+        ),
+    ),
+    (
+        "embedded_system_tag",
+        re.compile(
+            r"(?i)(<\s*system\s*>|<<\s*SYS\s*>>|\[INST\].*ignore)",
+        ),
+    ),
+]
